+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### 2.2.0 Alpha 1 (XXXX-YY-ZZ)
 
+* instruction file name for most PIPs has default value provided by `wcf\system\package\plugin\IPackageInstallationPlugin::getDefaultFilename()`.
 * `options` support for cronjobs.
+* `name` attribute for cronjob PIP (`cronjobName` for cronjob objects).
 * `permissions` and `options` support for event listeners.
+* `name` attribute for event listener PIP (`listenerName` for event listener objects).
 * `permissions` and `options` support for template listeners.
 * `wcf\data\TDatabaseObjectOptions` and `wcf\data\TDatabaseObjectPermissions` for database object-bound options and permissions validation.
 * `wcf\system\cache\builder\EventListenerCacheBuilder` returns `wcf\data\event\listener\EventListener` objects instead of data arrays.